Fix bug that crashes daemon when client disconnects
If the client closes the TCP connection while the server is sending
data, we crash like this:

[pid 22887] sendfile(4, 7, NULL, 1861208560) = -1 EPIPE (Broken pipe)
[pid 22887] --- SIGPIPE {si_signo=SIGPIPE, si_code=SI_USER, si_pid=22879, si_uid=65535} ---
[pid 22887] close(7)                    = 0
[pid 22887] write(4, "HTTP/1.1 500 Internal Server Err"..., 38) = -1 EPIPE (Broken pipe)
[pid 22887] --- SIGPIPE {si_signo=SIGPIPE, si_code=SI_USER, si_pid=22879, si_uid=65535} ---
[pid 22887] --- SIGILL {si_signo=SIGILL, si_code=ILL_ILLOPN, si_addr=0x400172} ---
[pid 22887] +++ killed by SIGILL +++
[pid 22888] +++ killed by SIGILL +++
+++ killed by SIGILL +++

The real issue here is that -EPIPE isn't recognized as an indication
that the client went away, but the crash happens because of how I
implemented the error-handler macro. I used "0" as a shorthand for
"no error handler", but the macro didn't account for this, so it was
literally assembling to 'jmp $0'.

Something about this confused the linker (or there's some deep magic
here I just don't understand), because it ended up doing this:

000000000040015c <__die>:
  40015c:	b8 01 00 00 00       	mov    $0x1,%eax
  400161:	48 89 df             	mov    %rbx,%rdi
  400164:	48 89 ce             	mov    %rcx,%rsi
  400167:	48 89 d2             	mov    %rdx,%rdx
  40016a:	0f 05                	syscall
  40016c:	48 09 c0             	or     %rax,%rax
  40016f:	0f 88 fc ff ff ff    	js     400171 <__die+0x15>
  400175:	b8 03 00 00 00       	mov    $0x3,%eax

The jump goes to 0x400171, which is actually in the middle of the
next instruction. So that's how we ended up getting SIGILL out of
this whole mess.